requirements.txt
annotated-types==0.7.0
anyio==4.9.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asgiref==3.11.0
asttokens==3.0.1
async-lru==2.0.5
attrs==24.2.0
babel==2.17.0
beautifulsoup4==4.14.3
bleach==6.3.0
blinker==1.8.2
CacheControl==0.14.2
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
comm==0.2.3
contourpy==1.3.3
cryptography==44.0.0
cycler==0.12.1
debugpy==1.8.19
decorator==5.2.1
defusedxml==0.7.1
distlib==0.3.8
dj-database-url==2.3.0
Django==5.2.7
django-rest-framework==0.1.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
dnspython==2.7.0
email_validator==2.2.0
et_xmlfile==2.0.0
executing==2.2.1
fastapi==0.115.14
fastapi-cli==0.0.7
fastjsonschema==2.21.2
filelock==3.15.4
firebase-admin==6.6.0
fonttools==4.61.1
fqdn==1.5.1
gitdb==4.0.11
google-api-core==2.24.1
google-api-python-client==2.160.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.20.0
google-cloud-storage==3.0.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
greenlet==3.2.3
grpcio==1.70.0
grpcio-status==1.70.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
idna==3.10
ipykernel==7.1.0
ipython==9.8.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.8
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.3
json5==0.12.1
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.7.0
jupyter_core==5.9.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.5.1
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
jupyterlab_widgets==3.0.16
kiwisolver==1.4.9
lark==1.3.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.8
matplotlib-inline==0.2.1
mdurl==0.1.2
mistune==3.2.0
msgpack==1.1.0
mysql-connector-python==9.0.0
mysqlclient==2.2.7
narwhals==1.10.0
nbclient==0.10.4
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
notebook==7.5.1
notebook_shim==0.2.4
numpy==2.1.2
openpyxl==3.1.5
packaging==24.1
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.5
php==1.2.1
pillow==10.4.0
platformdirs==4.2.2
prometheus_client==0.23.1
prompt_toolkit==3.0.52
proto-plus==1.26.0
protobuf==5.28.3
psutil==7.1.3
psycopg2-binary==2.9.11
pure_eval==0.2.3
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.18.0
PyJWT==2.10.1
pymongo==4.15.1
PyMySQL==1.1.1
pyparsing==3.2.1
pypng==0.20220715.0
PyQRCode==1.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-json-logger==4.0.0
python-multipart==0.0.20
pytz==2024.2
pywinpty==3.0.2
PyYAML==6.0.2
pyzmq==27.1.0
qrcode==7.4.2
referencing==0.35.1
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==13.9.3
rich-toolkit==0.14.8
rpds-py==0.20.0
rsa==4.9
scikit-learn==1.8.0
scipy==1.16.3
segno==1.6.6
Send2Trash==1.8.3
setuptools==80.9.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.8.1
SQLAlchemy==2.0.41
sqlparse==0.5.0
stack-data==0.6.3
starlette==0.46.2
tenacity==9.0.0
terminado==0.18.1
threadpoolctl==3.6.0
tinycss2==1.4.0
toml==0.10.2
tornado==6.4.1
traitlets==5.14.3
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2024.1
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.34.2
virtualenv==20.26.3
watchdog==5.0.3
watchfiles==1.1.0
wcwidth==0.2.14
webcolors==25.10.0
webencodings==0.5.1
websocket-client==1.9.0
websockets==15.0.1
whitenoise==6.9.0
widgetsnbextension==4.0.15
